split out middleware
